tests/wpt/tests/css/css-values/container-progress-serialize.tentative.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-values-5/#container-progress-func">
<link rel="author" title="sakhapov@chromuim.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../support/serialize-testcommon.js"></script>
<div id="outer-container">
  <div id="inner-container">
    <div id=target></div>
  </div>
</div>
<style>
:root {
  font-size: 10px;
}
#outer-container {
  container: my-container / size;
  width: 322px;
  height: 228px;
}
#inner-container {
  container-type: size;
  width: 228px;
  height: 322px;
}
#target {
  font-size: 10px;
}
</style>
<script>
function test_serialization(t,s,c) {
    test_specified_serialization('opacity', t, s);
    test_specified_serialization('transform', `scale(${t})`, `scale(${s})`);
    test_computed_serialization('opacity', t, c);
    test_computed_serialization('transform', `scale(${t})`, `matrix(${c}, 0, 0, ${c}, 0, 0)`);
}

test_serialization(
    'calc(container-progress(width, 0px, 1px) / 1000)',
    'calc(0.001 * container-progress(width, 0px, 1px))',
    '0.228',
);
test_serialization(
    'calc(0.1 * container-progress(height of my-container, 0px, 10em))',
    'calc(0.1 * container-progress(height of my-container, 0px, 10em))',
    '0.228',
);
</script>
